+++ src/evt/onWindowLoad/index.js
@@ -4,17 +4,45 @@ import { getDocument, getWindow } from "ssr-window";
 /**
  * Runs callback when page has fully loaded
  * @param cb{Function} - callback function
+ * @param isAutoInit{Boolean=} - attaches event immediately
  * @see https://developer.mozilla.org/en-US/docs/Web/API/Window/load_event
+ * @return {{ handler: Function, addListener: Function, removeListener: Function }}
  * @example
  * // How to detect when whole page has loaded?
  * const callback = () => console.log("Page loaded");
  * onWindowLoad(callback);
  * */
-const onWindowLoad = (cb) => {
+const onWindowLoad = (cb, isAutoInit = true) => {
 
   ow(cb, ow.function);
+  ow(isAutoInit, ow.optional.boolean);
 
-  getDocument().readyState === "complete" ? cb() : getWindow().addEventListener("load", cb);
+  const handler = () => {
+    cb();
+  };
+
+  const addListener = () => {
+    if (getDocument().readyState === "complete") {
+      handler();
+    } else {
+      getWindow().addEventListener("load", handler);
+    }
+  };
+
+  const removeListener = () => {
+    getWindow().removeEventListener("load", handler);
+  };
+
+  if (isAutoInit) {
+    addListener();
+  }
+
+  return {
+    handler,
+    addListener,
+    removeListener
+  };
+
 };
 
 export {